pyproject.toml
[project]
name = "cloud-finops-env"
version = "1.0.0"
description = "Cloud FinOps Cost Optimization Sandbox — OpenEnv environment"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openenv-core>=0.2.1",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "openai>=1.51.0",
    "httpx>=0.27.0,<0.28.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.wheel]
packages = ["server"]